Lectures 1 - 4

+++ app/Http/Controllers/CategoryController.php
@@ -13,10 +13,8 @@ class CategoryController extends Controller
      */
     public function index()
     {
-
-
-     $categories=   Category::all();
-        return view('categories.index',compact('categories'));
+     $categories=   Category::paginate(10);
+        return view('categories.index')->with('categories',$categories);
     }
 
     /**
@@ -61,7 +59,7 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
-        //
+        return view('categories.edit',compact('category'));
     }
 
     /**
@@ -69,7 +67,16 @@ class CategoryController extends Controller
      */
     public function update(UpdateCategoryRequest $request, Category $category)
     {
-        //
+        $category->update(
+            [
+                'name'=>$request->name,
+                "description"=>$request->description,
+                "type"=>$request->type,
+                "status"=>isset($request->status)
+            ]
+        );
+        toastr()->success('category was updated');
+        return redirect(route('categories.index')) ;
     }
 
     /**
@@ -77,7 +84,13 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $category->delete();
+        if($category->id>4) {
+            $category->delete();
+            // Display an error toast with no title
+            toastr()->success('تم الحذف بنجاح');
+        }
+        else
+            toastr()->error('لايمكن حذف هذا الصنف');
         return redirect()->back();
     }
 }
